+++ services/ordering/src/Ordering.Application/UseCases/CreateOrder/ICreateOrderUseCase.cs
@@ -3,7 +3,7 @@ namespace Choice.Ordering.Application.UseCases.CreateOrder
 {
     public interface ICreateOrderUseCase
     {
-        Task Execute(string receiverId, int orderRequestId, double price, double prepayment, int deadline,
+        Task Execute(string receiverId, int orderRequestId, int price, int prepayment, int deadline,
             DateTime enrollmentDate);
 
         void SetOutputPort(IOutputPort outputPort);
